add components for T2
