+++ react-init/src/components/news.js
@@ -1,0 +1,21 @@
+import React from 'react';
+
+import '@/assets/css/news.css'
+
+class News extends React.Component {
+  render() {
+    return (<div id="news">
+      <div id="news_background">
+        <div id="news_content">
+          <p>FALL & WINTER</p>
+          <button>Shop Now</button>
+        </div>
+      </div>
+      <div id="news_free_shipping_worlwide">
+        <p>FREE SHIPPING WORLWIDE</p>
+      </div>
+    </div>)
+  }
+}
+
+export default News
