+++ react-init/src/index.js
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 import React from 'react';
-import T2 from './T2'
+import T2 from '@/pages/t2'
 
 ReactDOM.render(
   <T2 />,
